+++ lib/utils.h
@@ -15,6 +15,6 @@ float relu_derivative(float x);
 float update_output(float input, float weight, float bias);
 
 float gradient_descent(float x, float rate, float dx);
-double generate_random(void);
+float generate_random(void);
 
 #endif
